split qwen36 moe shared decode types
