+++ crates/runner/src/main.rs
@@ -40,6 +40,7 @@ mod qwen36_moe_speculative;
 mod qwen36_moe_state;
 mod qwen36_moe_telemetry;
 mod qwen36_moe_timing;
+mod qwen36_moe_types;
 mod qwen36_moe_vmm;
 mod registry;
 mod specprefill;
