+++ crates/runner/src/qwen36_moe_session.rs
@@ -3,9 +3,9 @@ use gpu_hal::{GpuBuffer, ScalarType};
 use model_store::BakedStore;
 use qwen36_moe::config::TextConfig;
 
-use crate::qwen36_moe_decode::{LayerBuffers, MtpLayerBuffers, MultiLayerGeom};
 use crate::qwen36_moe_host::{host_load_bytes, load_lm_head_bf16};
 use crate::qwen36_moe_layers::load_to_gpu;
+use crate::qwen36_moe_types::{LayerBuffers, MtpLayerBuffers, MultiLayerGeom};
 
 const MIB: f64 = (1024 * 1024) as f64;
 const QWEN36_NUM_SPECULATIVE_TOKENS: usize = 3;
