+++ crates/runner/src/qwen36_moe_vmm.rs
@@ -3,10 +3,10 @@ use gpu_hal::{Backend, VirtualArena};
 use model_store::BakedStore;
 use qwen36_moe::config::TextConfig;
 
-use crate::qwen36_moe_decode::{AttnLayerBuffers, LayerBuffers, MultiLayerGeom};
 use crate::qwen36_moe_layers::{load_all_layer_buffers, Qwen36WeightMode};
 use crate::qwen36_moe_residency::{MoeExpertResidencyConfig, MoeExpertResidencyManager};
 use crate::qwen36_moe_telemetry::{MoeIslandPrefetchMode, MoeSparseTelemetry, VirtualKvStats};
+use crate::qwen36_moe_types::{AttnLayerBuffers, LayerBuffers, MultiLayerGeom};
 
 const MIB: f64 = (1024 * 1024) as f64;
 pub(crate) const DEFAULT_SPARSE_MOE_PREFETCH_RANKS: usize = 4;
